requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.0 